# Migrate log_sigmoid (forward and backward) to ATen (CUDA)