--- a/BUILD.bazel
+++ b/BUILD.bazel
@@ -408,7 +408,6 @@ filegroup(
         "aten/src/THCUNN/ELU.cu.cc",
         "aten/src/THCUNN/HardTanh.cu.cc",
         "aten/src/THCUNN/LeakyReLU.cu.cc",
-        "aten/src/THCUNN/LogSigmoid.cu.cc",
         "aten/src/THCUNN/MultiLabelMarginCriterion.cu.cc",
         "aten/src/THCUNN/MultiMarginCriterion.cu.cc",
         "aten/src/THCUNN/SoftMarginCriterion.cu.cc",