--- a/aten/src/THCUNN/CMakeLists.txt
+++ b/aten/src/THCUNN/CMakeLists.txt
@@ -1,5 +1,4 @@
 set(ATen_CUDA_SRCS ${ATen_CUDA_SRCS}
-${CMAKE_CURRENT_SOURCE_DIR}/LogSigmoid.cu
 ${CMAKE_CURRENT_SOURCE_DIR}/MultiLabelMarginCriterion.cu
 ${CMAKE_CURRENT_SOURCE_DIR}/MultiMarginCriterion.cu
 ${CMAKE_CURRENT_SOURCE_DIR}/SpatialClassNLLCriterion.cu