--- a/aten/src/THCUNN/generic/THCUNN.h
+++ b/aten/src/THCUNN/generic/THCUNN.h
@@ -5,19 +5,6 @@
 #include <ATen/core/Reduction.h>
 #include <ATen/Generator.h>
 
-TORCH_CUDA_CU_API void THNN_(LogSigmoid_updateOutput)(
-    THCState* state,
-    THCTensor* input,
-    THCTensor* output,
-    THCTensor* buffer);
-
-TORCH_CUDA_CU_API void THNN_(LogSigmoid_updateGradInput)(
-    THCState* state,
-    THCTensor* input,
-    THCTensor* gradOutput,
-    THCTensor* gradInput,
-    THCTensor* buffer);
-
 TORCH_CUDA_CU_API void THNN_(MultiLabelMarginCriterion_updateOutput)(
     THCState* state,
     THCTensor* input,