… aten/src/ATen/cuda/LegacyTHFunctionsCUDA.cpp
… aten/src/THCUNN/LogSigmoid.cu
… aten/src/THCUNN/generic/LogSigmoid.cu