@ngimel 2021-07-06
It would make sense to add `logsigmoid` to OpInfo (and, since it is UnaryUfunc, test it with UnaryUfunc tests), but if it doesn't work with reasonable amount of effort we can land as is (just remove stubs), and fix whatever is failing later.